pyproject.toml
[project]
name = "textures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
]
